fix projects
