+++ src/components/project/projectsArray.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+const projectsArray = JSON.parse(localStorage.getItem('Projects')) || [];
+
+// const exportProjectsArray = () => projectsArray;
+
+export default { projectsArray };
